api/exec: implement ExecStartOpts
Currently this crate has no way to specify whether a started exec should be attached or
detached. Rectify this by implementing ExecStartOpts [0].

NB: This is technically a breaking change since I had to change the type signature of
create_and_exec() to accomodate the new options. If anyone has alternative suggestions on
how to handle this, let me know.

[0]: https://docs.docker.com/engine/api/v1.41/#operation/ExecStart

Signed-off-by: William Findlay <will@isovalent.com>